+++ src/de/omegazirkel/risingworld/adminutils/exports/Vector3Export.java
@@ -1,0 +1,4 @@
+package de.omegazirkel.risingworld.adminutils.exports;
+
+public record Vector3Export(float x, float y, float z) {
+}